changes made as per comments
